code/lamda-flask/requirements.txt
aws==0.2.5
awsebcli @ git+https://github.com/aws/aws-elastic-beanstalk-cli@63f6107ddd10884404c7dc3b604e20eca67fb219
bcrypt==4.1.2
blessed==1.20.0
blinker==1.7.0
boto==2.49.0
boto3==1.26.109
botocore==1.29.109
cement==2.8.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
cryptography==43.0.0.dev1
decorator==5.1.1
Deprecated==1.2.14
fabric==3.2.2
Flask==2.3.0
idna==3.6
iniconfig==2.0.0
invoke==2.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.4
packaging==23.2
paramiko==3.4.0
pathspec==0.10.1
pluggy==1.4.0
prettytable==3.9.0
pycparser==2.21
PyNaCl==1.5.0
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.6.2
semantic-version==2.8.5
six==1.16.0
termcolor==1.1.0
urllib3==1.26.18
wcwidth==0.1.9
Werkzeug==3.0.1
wrapt==1.16.0
